Update strings.xml (#9)
Clean-up of the French translation. I'm not very happy with my suggestion for "nothing but crickets" but the literal translation made absolutely no sense in French. Mine is not hitting the same colloquial friendly vibes but at least it's understandable.

Some lines in the app are not present in this file. I searched how to contribute to the translation to translate "About" ("À propos") and "History" ("Historique") but I've not found them in this file…